Editor: Extract LinkContainer from FormatToolbar (#9192)
* Editor: Extract LinkContainer from FormatToolbar

* Format toolbar: Improve derived state handling

* Rich Text: Remove obsolete DEFAULT_CONTROLS constant
